+++ js-webshim/dev/shims/picture.js
@@ -7,67 +7,26 @@
 			return document.createElement('img');
 		};
 	}
+
+	webshim.isReady('picture', true);
 	setTimeout(function(){
-		if(window.picturefill){
-			var sel = 'picture, img[srcset]';
-			webshims.addReady(function(context){
-				if(context.querySelector(sel)){
-					window.picturefill();
-				}
-			});
-		}
+		webshim.ready('matchMedia', function(){
+			if(window.picturefill){
+				var sel = 'picture, img[srcset]';
+				webshim.addReady(function(context){
+					if(context.querySelector(sel)){
+						window.picturefill();
+					}
+				});
+			}
+		});
 	});
 })();
 
 /*! Picturefill - v2.1.0-beta - 2014-06-03
  * http://scottjehl.github.io/picturefill
  * Copyright (c) 2014 https://github.com/scottjehl/picturefill/blob/master/Authors.txt; Licensed MIT */
-/*! matchMedia() polyfill - Test a CSS media type/query in JS. Authors & copyright (c) 2012: Scott Jehl, Paul Irish, Nicholas Zakas, David Knight. Dual MIT/BSD license */
 
-window.matchMedia || (window.matchMedia = function() {
-	"use strict";
-
-	// For browsers that support matchMedium api such as IE 9 and webkit
-	var styleMedia = (window.styleMedia || window.media);
-
-	// For those that don't support matchMedium
-	if (!styleMedia) {
-		var style       = document.createElement('style'),
-			script      = document.getElementsByTagName('script')[0],
-			info        = null;
-
-		style.type  = 'text/css';
-		style.id    = 'matchmediajs-test';
-
-		script.parentNode.insertBefore(style, script);
-
-		// 'style.currentStyle' is used by IE <= 8 and 'window.getComputedStyle' for all other browsers
-		info = ('getComputedStyle' in window) && window.getComputedStyle(style, null) || style.currentStyle;
-
-		styleMedia = {
-			matchMedium: function(media) {
-				var text = '@media ' + media + '{ #matchmediajs-test { width: 1px; } }';
-
-				// 'style.styleSheet' is used by IE <= 8 and 'style.textContent' for all other browsers
-				if (style.styleSheet) {
-					style.styleSheet.cssText = text;
-				} else {
-					style.textContent = text;
-				}
-
-				// Test if media query is true or false
-				return info.width === '1px';
-			}
-		};
-	}
-
-	return function(media) {
-		return {
-			matches: styleMedia.matchMedium(media || 'all'),
-			media: media || 'all'
-		};
-	};
-}());
 /*! Picturefill - Responsive Images that work today.
  *  Author: Scott Jehl, Filament Group, 2012 ( new proposal implemented by Shawn Jansepar )
  *  License: MIT/GPLv2
@@ -637,7 +596,7 @@ window.matchMedia || (window.matchMedia = function() {
 		}
 	}
 
-	runPicturefill();
+	webshim.ready('matchMedia', runPicturefill);
 
 	/* expose methods for testing */
 	picturefill._ = pf;